switch app, header, and layout to tsx and fix structure
